+++ src/lib/terkin/datalogger.py
@@ -18,17 +18,15 @@ from terkin.network import SystemWiFiMetrics
 from terkin.sensor import SensorManager, AbstractSensor
 from terkin.model import SensorReading, DataFrame
 from terkin.sensor.system import SystemMemoryFree, SystemTemperature, SystemVoltage, SystemUptime
-from terkin.driver.si7021_sensor import SI7021Sensor
-from terkin.driver.hx711_sensor import HX711Sensor
-from terkin.driver.max17043_sensor import MAX17043Sensor
-from terkin.driver.bme280_sensor import BME280Sensor
-from terkin.util import gc_disabled, ddformat
+from terkin.util import gc_disabled
 
 log = logging.getLogger(__name__)
 
 
 class TransientStorage:
-    """ """
+    """
+    Store last sensor reading for serving through HTTP API.
+    """
 
     def __init__(self):
         self.last_reading = {}
@@ -104,7 +102,9 @@ class TerkinDatalogger:
 
         # Turn off LTE modem and Bluetooth as we don't use them yet.
         # TODO: Make this configurable.
-        if self.application_info.platform_info.device_name in ['GPy', 'FiPy']:
+        has_lte_modem = self.application_info.platform_info.device_name in ['GPy', 'FiPy']
+        lte_enabled = self.settings.get('networking.lte.enabled')
+        if has_lte_modem and not lte_enabled:
             self.device.power_off_lte_modem()
 
         self.device.power_off_bluetooth()
@@ -135,6 +135,10 @@ class TerkinDatalogger:
 
         # Todo: Signal readyness by publishing information about the device (Microhomie).
         # e.g. ``self.device.publish_properties()``
+
+        self.setup_sensors()
+
+    def setup_sensors(self):
 
         # Setup sensors.
         log.info('Setting up sensors')
@@ -171,7 +175,7 @@ class TerkinDatalogger:
         # Run downstream mainloop handlers.
         self.duty_cycle()
 
-        # Sleep how ever.
+        # Sleep how configured
         self.sleep()
 
     def duty_cycle(self):
@@ -211,45 +215,52 @@ class TerkinDatalogger:
         machine.idle()
 
     def sleep(self):
-        """Sleep until the next measurement cycle."""
+        """Sleep or shutoff until the next measurement cycle."""
 
         lightsleep = self.settings.get('main.lightsleep', False)
         deepsleep = self.settings.get('main.deepsleep', False)
+        shutoff = self.settings.get('main.shutoff', False)
         interval = self.get_sleep_time()
 
         # Amend deep sleep intent when masked through maintenance mode.
         if self.device.status.maintenance is True:
             lightsleep = False
             deepsleep = False
+            shutoff = False
             log.info('Device is in maintenance mode. Skipping deep sleep and '
                      'adjusting sleep time to {} seconds.'.format(interval))
 
         # Prepare device shutdown.
         try:
-
             # Shut down sensor peripherals.
             self.sensor_manager.power_off()
 
             # Shut down networking.
-            if deepsleep:
+            if deepsleep or shutoff:
                 self.device.networking.stop()
 
         except Exception as ex:
             log.exc(ex, 'Power off failed')
 
-        # Activate device sleep mode.
-        try:
-            self.device.hibernate(interval, lightsleep=lightsleep, deepsleep=deepsleep)
+        if shutoff:
+            # shut off the MCU via DS3231
+            self.shutoff()
+        else:
+            # Activate device sleep mode.
+            try:
+                self.device.hibernate(interval, lightsleep=lightsleep, deepsleep=deepsleep)
 
-        # When hibernation fails, fall back to regular "time.sleep".
-        except Exception as ex:
-            log.exc(ex, 'Failed to hibernate, falling back to regular sleep')
-            # Todo: Emit error message here.
-            log.info('Sleeping for {} seconds'.format(interval))
-            time.sleep(interval)
+            # When hibernation fails, fall back to regular "time.sleep".
+            except Exception as ex:
+                log.exc(ex, 'Failed to hibernate, falling back to regular sleep')
+                # Todo: Emit error message here.
+                log.info('Sleeping for {} seconds'.format(interval))
+                time.sleep(interval)
 
     def get_sleep_time(self):
-        """ """
+        """ 
+        Calculate the next sleep intervall.
+        """
 
         interval = self.settings.get('main.interval', 60.0)
 
@@ -336,7 +347,9 @@ class TerkinDatalogger:
             #self.device.run_gc()
 
     def register_sensor(self, sensor_info):
-
+        """
+        Register one sensor.
+        """
         sensor_type = sensor_info.get('type', 'unknown').lower()
         sensor_id = sensor_info.get('id', sensor_info.get('key', sensor_type))
         description = sensor_info.get('description')
@@ -362,8 +375,10 @@ class TerkinDatalogger:
             sensor_address = None
 
         # Report sensor registration to user.
-        message = 'Setting up sensor with id={} and type={} on bus={} with address={} ' \
-                  'described as "{}"'.format(sensor_id, sensor_type, sensor_bus_name, sensor_address, description)
+        message = 'Setting up sensor with id={} and type={} on bus={} with address={}'.format(
+            sensor_id, sensor_type, sensor_bus_name, sensor_address)
+        if description:
+            message += ' described as "{}"'.format(description)
         log.info(message)
 
         # Backward compat.
@@ -450,49 +465,6 @@ class TerkinDatalogger:
             self.start_buttons()
             return
 
-        # Setup and register HX711 sensors.
-        elif sensor_type == 'hx711':
-            sensor_object = HX711Sensor(settings=sensor_info)
-            sensor_object.set_address(sensor_info.get('number', sensor_info.get('address', 0)))
-            sensor_object.register_pin('dout', sensor_info['pin_dout'])
-            sensor_object.register_pin('pdsck', sensor_info['pin_pdsck'])
-            sensor_object.register_parameter('scale', float(sensor_info['scale']))
-            sensor_object.register_parameter('offset', float(sensor_info['offset']))
-            sensor_object.register_parameter('gain', sensor_info.get('gain', 128))
-
-            # Select driver module. Use "gerber" (vanilla) or "heisenberg" (extended).
-            # hx711_sensor.select_driver('gerber')
-            sensor_object.select_driver('heisenberg')
-
-        # Setup and register SI7021 sensors.
-        elif sensor_type == 'si7021':
-
-            sensor_object = SI7021Sensor(settings=sensor_info)
-            if 'address' in sensor_info:
-                sensor_object.set_address(sensor_info['address'])
-            sensor_object.acquire_bus(sensor_bus)
-
-        elif sensor_type == 'max17043':
-
-            sensor_object = MAX17043Sensor(settings=sensor_info)
-            if 'address' in sensor_info:
-                sensor_object.set_address(sensor_info['address'])
-            sensor_object.acquire_bus(sensor_bus)
-
-        elif sensor_type == 'vedirect':
-
-            from terkin.driver.vedirect_sensor import VEDirectSensor
-            sensor_object = VEDirectSensor(settings=sensor_info)
-
-
-        elif sensor_type == 'ads1x15':
-
-            from terkin.driver.ads1x15_sensor import ADS1x15Sensor
-            sensor_object = ADS1x15Sensor(settings=sensor_info)
-            # if 'address' in sensor_info:
-            #    sensor_object.set_address(sensor_info['address'])
-            sensor_object.acquire_bus(sensor_bus)
-
         else:
             raise SensorUnknownError('Unknown sensor type "{}"'.format(sensor_type))
 
@@ -575,6 +547,7 @@ class TerkinDatalogger:
         # Debugging: Print sensor data before running telemetry.
         prettify_log = self.settings.get('sensors.prettify_log', False)
         if prettify_log:
+            from terkin.util import ddformat
             log.info('Sensor data:\n\n%s', ddformat(richdata, indent=11))
         else:
             log.info('Sensor data:  %s', data)
@@ -637,6 +610,7 @@ class TerkinDatalogger:
         return success
 
     def start_buttons(self):
+
         """
         Configure ESP32 touchpads.
         """
@@ -680,3 +654,77 @@ class TerkinDatalogger:
         # Will yield ``ValueError: Touch pad error``.
         #self.button_manager.setup_touchpad('P20', name='Touch8', location='Module-Right-Top-7th')
         #self.button_manager.setup_touchpad('P19', name='Touch9', location='Module-Right-Top-8th')
+
+    def shutoff(self):
+        """ shut off the MCU """
+
+        import DS3231tokei
+        import utime
+        from machine import Pin, RTC
+
+        # LED an
+        led = Pin(14,Pin.OUT)
+        horn = Pin(26,Pin.OUT)
+        led.value(1)
+        horn.value(1)
+        utime.sleep(1)
+        led.value(0)
+        horn.value(0)
+
+        bus = self.sensor_manager.get_bus_by_name('i2c:0')
+        ds = DS3231tokei.DS3231(bus.adapter)
+        interval = self.settings.get('main.interval.shutoff', 10) * 60  # convert from minutes to seconds
+        (year,month,day,dotw,hour,minute,second) = ds.getDateTime() # get the current time
+
+        print('Time is: ', day,hour,minute)
+
+        rtc = RTC() # create RTC
+        if year < 2001:
+            year = 2001 # sanity check, as of mpy 1.12 year must be >= 2001
+        rtc.init((year,month,day,dotw,hour,minute,second,0)) # set time
+
+        # Compute sleeping duration from measurement interval and elapsed time.
+        elapsed = int(self.duty_chrono.read())
+        now_secs = utime.mktime(utime.localtime())
+        wake_at = now_secs - elapsed + interval
+        if (wake_at - now_secs) < 180:  # don't shutoff for less than 3 minutes
+            wake_at += interval
+
+        print('Now:',now_secs, 'Wake at:', wake_at)
+
+        (year,month,day,hour,minute,second, dotw, doty) = utime.localtime(wake_at) # convert the wake up time
+
+        # set alarm
+        ds.setAlarm2(day,hour,minute, DS3231tokei.A2_ON_HOUR_MINUTE)
+
+        print('Wake at: ', day,hour,minute)
+
+        # turn off MCU via MOSFET
+        print('Good night')
+
+        utime.sleep(1)
+
+        ds.enableAlarm2()
+        ds.resetAlarm2()        
+
+        # The End
+
+    def scale_wizard(self):
+        """
+        Invoke scale adjustment wizard.
+
+        Synopsis:
+
+        - Invoke Terkin datalogger.
+        - Interrupt by pressing CTRL+C.
+        - Type ``datalogger.scale_wizard()``.
+        """
+
+        # Setup sensors.
+        self.setup_sensors()
+
+        # Invoke scale adjustment routine.
+        from terkin.sensor.scale import ScaleAdjustment
+        adj = ScaleAdjustment(sensor_manager=self.sensor_manager)
+        adj.start_wizard()
+