+++ src/lib/terkin/network/lora.py
@@ -97,30 +97,32 @@ class LoRaDriverPycom:
                 except:
                     pass
 
-        import binascii
+        if not self.lora.has_joined():
+            import binascii
 
-        # Over-the-Air Activation (OTAA)
-        if self.settings.get('networking.lora.activation') == 'otaa':
-            app_eui = binascii.unhexlify(self.settings.get('networking.lora.otaa.application_eui'))
-            app_key = binascii.unhexlify(self.settings.get('networking.lora.otaa.application_key'))
+            datarate_join = self.settings.get('networking.lora.datarate_join')
 
-            if not self.lora.has_joined():
+            # Over-the-Air Activation (OTAA)
+            if self.settings.get('networking.lora.activation') == 'otaa':
+                app_eui = binascii.unhexlify(self.settings.get('networking.lora.otaa.application_eui'))
+                app_key = binascii.unhexlify(self.settings.get('networking.lora.otaa.application_key'))
+
                 log.info('[LoRa] Attaching to the LoRaWAN network using OTAA')
                 if self.settings.get('networking.lora.otaa.device_eui') is None:
-                    self.lora.join(activation=LoRa.OTAA, auth=(app_eui, app_key), timeout=0)
+                    self.lora.join(activation=LoRa.OTAA, auth=(app_eui, app_key), timeout=0, dr=datarate_join)
                 else:
                     dev_eui = binascii.unhexlify(self.settings.get('networking.lora.otaa.device_eui'))
-                    self.lora.join(activation=LoRa.OTAA, auth=(dev_eui, app_eui, app_key), timeout=0, dr=0)
+                    self.lora.join(activation=LoRa.OTAA, auth=(dev_eui, app_eui, app_key), timeout=0, dr=datarate_join)
 
-        # Activation by Personalization (ABP)
-        elif self.settings.get('networking.lora.activation') == 'abp':
-            log.info('[LoRa] Attaching to the LoRaWAN network using ABP')
-            import struct
-            dev_addr = struct.unpack(">l", binascii.unhexlify(self.settings.get('networking.lora.abp.device_address')))[0]
-            nwk_swkey = binascii.unhexlify(self.settings.get('networking.lora.abp.network_session_key'))
-            app_swkey = binascii.unhexlify(self.settings.get('networking.lora.abp.app_session_key'))
+            # Activation by Personalization (ABP)
+            elif self.settings.get('networking.lora.activation') == 'abp':
+                import struct
+                dev_addr = struct.unpack(">l", binascii.unhexlify(self.settings.get('networking.lora.abp.device_address')))[0]
+                nwk_swkey = binascii.unhexlify(self.settings.get('networking.lora.abp.network_session_key'))
+                app_swkey = binascii.unhexlify(self.settings.get('networking.lora.abp.app_session_key'))
 
-            self.lora.join(activation=LoRa.ABP, auth=(dev_addr, nwk_swkey, app_swkey), timeout=0, dr=0)
+                log.info('[LoRa] Attaching to the LoRaWAN network using ABP')
+                self.lora.join(activation=LoRa.ABP, auth=(dev_addr, nwk_swkey, app_swkey), timeout=0, dr=datarate_join)
 
     def ensure_connectivity(self):